Add footer back to [slug] pages
I still think it's a little awkward to have "MADE BY PURDUE HACKERS" at the bottom of every page, but I think for now this is better because otherwise the RSVP form is stuck at the bottom of the page and it feels awkward. The copy can be changed later if needed.