fix: use UTC time for constructing GCal date and time
This ensures that the time for the event is correct regardless of the
user's time zone setting.

Looking at the code it looks weird that we're converting from UTC to
zoned time and then back to UTC. However this is intentional as date
objects in JS have no TZ value. This is even showcased in the official
documentation, see: https://www.npmjs.com/package/date-fns-tz#format